Fix some compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/3905



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@16760 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/cintltst/cmsgtst.c
@@ -61,12 +61,12 @@ static void InitStrings( void )
         return;
 
     for (i=0; i < cnt_testCases; i++ ) {
-        uint32_t strSize = strlen(txt_testCasePatterns[i]) + 1;
+        uint32_t strSize = (uint32_t)strlen(txt_testCasePatterns[i]) + 1;
         testCasePatterns[i]=(UChar*)malloc(sizeof(UChar) * strSize);
         u_uastrncpy(testCasePatterns[i], txt_testCasePatterns[i], strSize);
     }
     for (i=0; i < cnt_testCases; i++ ) {
-        uint32_t strSize = strlen(txt_testResultStrings[i]) + 1;
+        uint32_t strSize = (uint32_t)strlen(txt_testResultStrings[i]) + 1;
         testResultStrings[i] = (UChar*)malloc(sizeof(UChar) * strSize);
         u_uastrncpy(testResultStrings[i], txt_testResultStrings[i], strSize);
     }
